r.volume: change int type to prevent overflow, see https://trac.osgeo.org/grass/ticket/3004 (merge from trunk, https://trac.osgeo.org/grass/changeset/68376)
git-svn-id: https://svn.osgeo.org/grass/grass/branches/releasebranch_7_0@68378 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ raster/r.volume/centroids.c
@@ -22,7 +22,7 @@
 int centroids(int fd,		/* File descriptor of map layer to process */
 	      /* This file is assumed to be opened before calling */
 	      /*   centroids. */
-	      int *e, int *n,	/* Pointers to arrays at least max+1 long */
+	      unsigned long *e, unsigned long *n,	/* Pointers to arrays at least max+1 long */
 	      int method,	/* 0, 1, or 2; see above. */
 	      int max)
 {				/* Highest positive cat number in map layer */
